Add files via upload
I made this file on Windows since that is the main OS that I run and I have a virtual box running Ubuntu that runs my MySQL databases. My plan is to use a .txt file to transfer the data that I made on Windows to Ubuntu so I can insert the data that I generated into my database.

DISCLAIMER: This file is not polished, yet.